Refactored part of the client side action logging to make Youtube actions actually working
The callback which is called by the JS API of a loaded Youtube video must be defined before the video loads, this required separating the connected methods into its own file for inclusion in the 'header' scope. Then the whole client side queue mechanism had to be improved so that the queue can be populated with Youtube actions that need processing prior to having the Nitro library fully loaded.